cm/cm_dhexppot/requirements.txt
shapely==1.7.1
rasterio==1.1.8
pyproj==3.0.0.post1
geopandas
urllib3==1.26.5
jenkspy==0.2.0
numpy==1.19.5
scipy==1.6.3
rasterstats==0.14.0
